requirements.txt
streamlit
Pillow
pillow-heif